requirements.txt
Pillow==10.0.0
PyInstaller==6.1.0
